+++ agent/functional_tests/tests/functionaltests_test.go
@@ -26,16 +26,17 @@ import (
 	"testing"
 	"time"
 
+	"github.com/aws/amazon-ecs-agent/agent/ec2"
 	ecsapi "github.com/aws/amazon-ecs-agent/agent/ecs_client/model/ecs"
 	. "github.com/aws/amazon-ecs-agent/agent/functional_tests/util"
-	docker "github.com/fsouza/go-dockerclient"
-
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/awserr"
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/cloudwatch"
 	"github.com/aws/aws-sdk-go/service/cloudwatchlogs"
+	ec2sdk "github.com/aws/aws-sdk-go/service/ec2"
 	"github.com/docker/docker/pkg/system"
+	docker "github.com/fsouza/go-dockerclient"
 	"github.com/pborman/uuid"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -632,4 +633,83 @@ func testV3TaskEndpoint(t *testing.T, taskName, containerName, networkMode, awsl
 
 	exitCode, _ := task.ContainerExitcode(containerName)
 	assert.Equal(t, 42, exitCode, fmt.Sprintf("Expected exit code of 42; got %d", exitCode))
+}
+
+func TestContainerInstanceTags(t *testing.T) {
+	// We need long container instance ARN for tagging APIs, PutAccountSettingInput
+	// will enable long container instance ARN.
+	putAccountSettingInput := ecsapi.PutAccountSettingInput{
+		Name:  aws.String("containerInstanceLongArnFormat"),
+		Value: aws.String("enabled"),
+	}
+	_, err := ECS.PutAccountSetting(&putAccountSettingInput)
+	assert.NoError(t, err)
+
+	ec2Key := "ec2Key"
+	ec2Value := "ec2Value"
+	localKey := "localKey"
+	localValue := "localValue"
+	commonKey := "commonKey"
+	commonKeyEC2Value := "commonEC2Value"
+	commonKeyLocalValue := "commonKeyLocalValue"
+
+	// Get instance ID.
+	ec2MetadataClient := ec2.NewEC2MetadataClient(nil)
+	instanceID, err := ec2MetadataClient.InstanceID()
+	assert.NoError(t, err)
+
+	// Add tags to the instance.
+	ec2Client := ec2.NewClientImpl(*ECS.Config.Region)
+	createTagsInput := ec2sdk.CreateTagsInput{
+		Resources: []*string{aws.String(instanceID)},
+		Tags: []*ec2sdk.Tag{
+			{
+				Key:   aws.String(ec2Key),
+				Value: aws.String(ec2Value),
+			},
+			{
+				Key:   aws.String(commonKey),
+				Value: aws.String(commonKeyEC2Value),
+			},
+		},
+	}
+	_, err = ec2Client.CreateTags(&createTagsInput)
+	assert.NoError(t, err)
+
+	// Set the env var for tags and start Agent.
+	agentOptions := &AgentOptions{
+		ExtraEnvironment: map[string]string{
+			"ECS_CONTAINER_INSTANCE_PROPAGATE_TAGS_FROM": "ec2_instance",
+			"ECS_CONTAINER_INSTANCE_TAGS": fmt.Sprintf(`{"%s": "%s", "%s": "%s"}`,
+				localKey, localValue, commonKey, commonKeyLocalValue),
+		},
+	}
+	agent := RunAgent(t, agentOptions)
+	defer agent.Cleanup()
+	// Change the required Agent version to v1.22.0 during 1.22.0 staging or after staging.
+	agent.RequireVersion(">=1.21.0")
+
+	// Verify the tags are registered.
+	ListTagsForResourceInput := ecsapi.ListTagsForResourceInput{
+		ResourceArn: aws.String(agent.ContainerInstanceArn),
+	}
+	ListTagsForResourceOutput, err := ECS.ListTagsForResource(&ListTagsForResourceInput)
+	assert.NoError(t, err)
+	registeredTags := ListTagsForResourceOutput.Tags
+	expectedTagsMap := map[string]string{
+		ec2Key:    ec2Value,
+		localKey:  localValue,
+		commonKey: commonKeyLocalValue, // The value of common tag key should be local common tag value
+	}
+
+	// Here we only verify that the tags we've defined in the test are registered, because the
+	// test instance may have some other tags that are unknown to us.
+	for _, registeredTag := range registeredTags {
+		registeredTagKey := aws.StringValue(registeredTag.Key)
+		if expectedVal, ok := expectedTagsMap[registeredTagKey]; ok {
+			assert.Equal(t, expectedVal, aws.StringValue(registeredTag.Value))
+			delete(expectedTagsMap, registeredTagKey)
+		}
+	}
+	assert.Zero(t, len(expectedTagsMap))
 }