+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
+
 ## 1.22.0-dev
 * Feature - Add support for ECS Secrets integrating with AWS Systems Manager Parameter Store
+* Feature - Introduce two environment variables `ECS_CONTAINER_INSTANCE_PROPAGATE_TAGS_FROM` and `ECS_CONTAINER_INSTANCE_TAGS` to support ECS tagging [#1618](https://github.com/aws/amazon-ecs-agent/pull/1618)
 
 ## 1.21.0
 * Feature - Add v3 task metadata support for awsvpc, host and bridge network mode